requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.30.0
chromadb>=0.5.0
httpx>=0.27.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
docling>=2.0.0
PyMuPDF>=1.24.0
